Retrieve Covid19 Data

This functions hits the Johns Hopkins Covid Data Repository and filters down to North Carolina Data. Alternatively, it can retrieve any state.

Usage,

get_covid_state(
  state = "North Carolina",
  select_county = NULL,
  data_source = c("cone", "hopkins"),
  reporting_adj = FALSE
)

Arguments

state

with a default of North Carolina

select_county

the county, if desired

data_source

which data source you would like to use one of "cone" or "hopkins"

reporting_adj

a boolean, default of FALSE which adjust for two known issues with North Carolina Report occurring on 2020-09-25 when all antigen tests were added and on 2020-11-13 where only 10 hours of data were reported and the remaining cases were rolled into 2020-11-14. Also corrects for reporting lapses on Thanksgiving and Christmas Eve/ Christmas Day. On 2021-02-28 the dashboard was taken down for maintenance and thus unavailable on that day. Additionally note that on 2021-03-13 NCDHHS elected to not report data on Sundays. On 2021-03-19 NCDHHS notified users that on 2021-03-26 data would be updated Monday - Friday. Updated for Good Friday in which no results were posted.

Examples

if (FALSE) {
# To get all counties
get_covid_state()

# To get a single county
get_covid_state(select_county = "Guilford")
}