Pull Vaccinations by County and Demographic

Pulled from <conedatascience/covid-data>. The column vax_n returns vaccination counts. Population references are available for the user to calculate distribution and/or per capita rates.

get_vaccinations_demo(county_list = NULL, demographic = NULL, status = NULL)

Arguments

county_list
a string, the county to pull. default is all counties.

demographic
a string, the demographic requested. one or more of "race", "age", "ethnicity", "sex". default is all demographics

status
string, either 'partial' or 'full' for number of people with partially or fully vaccinated. Default of NULL returns both.

Examples

{

get_vaccinations_demo(county_list = "Guilford",
                      demographic = 'race', status = 'partial')

}

#>     date_pulled    week_of aggregation_level   county county_pop demographic
#>  1:  2021-04-05 2020-12-14            County Guilford     537174        Race
#>  2:  2021-04-05 2020-12-21            County Guilford     537174        Race
#>  3:  2021-04-05 2020-12-28            County Guilford     537174        Race
#>  4:  2021-04-05 2021-01-04            County Guilford     537174        Race
#>  5:  2021-04-05 2021-01-11            County Guilford     537174        Race
#>  6:  2021-04-05 2021-01-18            County Guilford     537174        Race
#>  7:  2021-04-05 2021-01-25            County Guilford     537174        Race
#>  8:  2021-04-05 2021-02-01            County Guilford     537174        Race
#>  9:  2021-04-05 2021-02-08            County Guilford     537174        Race
#> 10:  2021-04-05 2021-02-15            County Guilford     537174        Race
#> 11:  2021-04-05 2021-02-22            County Guilford     537174        Race
#> 12:  2021-04-05 2021-03-01            County Guilford     537174        Race
#> 13:  2021-04-05 2021-03-08            County Guilford     537174        Race
#> 14:  2021-04-05 2021-03-15            County Guilford     537174        Race
#> 15:  2021-04-05 2021-03-22            County Guilford     537174        Race
#> 16:  2021-04-05 2021-03-29            County Guilford     537174        Race
#> 17:  2021-04-05 2021-04-05            County Guilford     537174        Race
#> 18:  2021-04-05 2020-12-14            County Guilford     537174        Race
#> 19:  2021-04-05 2020-12-21            County Guilford     537174        Race
#> 20:  2021-04-05 2020-12-28            County Guilford     537174        Race
#> 21:  2021-04-05 2021-01-04            County Guilford     537174        Race
#> 22:  2021-04-05 2021-01-11            County Guilford     537174        Race
#> 23:  2021-04-05 2021-01-18            County Guilford     537174        Race
#> 24:  2021-04-05 2021-01-25            County Guilford     537174        Race
#> 25:  2021-04-05 2021-02-01            County Guilford     537174        Race
#> 26:  2021-04-05 2021-02-08            County Guilford     537174        Race
#> 27:  2021-04-05 2021-02-15            County Guilford     537174        Race
#> 28:  2021-04-05 2021-02-22            County Guilford     537174        Race
#> 29:  2021-04-05 2021-03-01            County Guilford     537174        Race
#> 30:  2021-04-05 2021-03-08            County Guilford     537174        Race
#> 31:  2021-04-05 2021-03-15            County Guilford     537174        Race
#> 32:  2021-04-05 2021-03-22            County Guilford     537174        Race
#> 33:  2021-04-05 2021-03-29            County Guilford     537174        Race
#> 34:  2021-04-05 2021-04-05            County Guilford     537174        Race
#> 35:  2021-04-05 2020-12-14            County Guilford     537174        Race
#> 36:  2021-04-05 2020-12-21            County Guilford     537174        Race
#> 37:  2021-04-05 2020-12-28            County Guilford     537174        Race
#> 38:  2021-04-05 2021-01-04            County Guilford     537174        Race
#> 39:  2021-04-05 2021-01-11            County Guilford     537174        Race
#> 40:  2021-04-05 2021-01-18            County Guilford     537174        Race
#> 41:  2021-04-05 2021-01-25            County Guilford     537174        Race
#> 42:  2021-04-05 2021-02-01            County Guilford     537174        Race
#> 43:  2021-04-05 2021-02-08            County Guilford     537174        Race
#> 44:  2021-04-05 2021-02-15            County Guilford     537174        Race
#> 45:  2021-04-05 2021-02-22            County Guilford     537174        Race
#> 46:  2021-04-05 2021-03-01            County Guilford     537174        Race
#> 47:  2021-04-05 2021-03-08            County Guilford     537174        Race
#> 48:  2021-04-05 2021-03-15            County Guilford     537174        Race
#> 49:  2021-04-05 2021-03-22            County Guilford     537174        Race
#> 50:  2021-04-05 2021-03-29            County Guilford     537174        Race
#> 51:  2021-04-05 2021-04-05            County Guilford     537174        Race
#> 52:  2021-04-05 2020-12-28            County Guilford     537174        Race
#> 53:  2021-04-05 2021-01-18            County Guilford     537174        Race
#> 54:  2021-04-05 2021-01-25            County Guilford     537174        Race
#> 55:  2021-04-05 2021-02-01            County Guilford     537174        Race
#> 56:  2021-04-05 2021-02-08            County Guilford     537174        Race
#> 57:  2021-04-05 2021-02-15            County Guilford     537174        Race
#> 58:  2021-04-05 2021-02-22            County Guilford     537174        Race
#> 59:  2021-04-05 2021-03-01            County Guilford     537174        Race
#> 60:  2021-04-05 2021-03-08            County Guilford     537174        Race
#> 61:  2021-04-05 2021-03-15            County Guilford     537174        Race
#> 62:  2021-04-05 2021-03-22            County Guilford     537174        Race
#> 63:  2021-04-05 2021-03-29            County Guilford     537174        Race
#> 64:  2021-04-05 2020-12-14            County Guilford     537174        Race
#> 65:  2021-04-05 2020-12-21            County Guilford     537174        Race
#> 66:  2021-04-05 2020-12-28            County Guilford     537174        Race
#> 67:  2021-04-05 2021-01-04            County Guilford     537174        Race
#> 68:  2021-04-05 2021-01-11            County Guilford     537174        Race
#> 69:  2021-04-05 2021-01-18            County Guilford     537174        Race
#> 70:  2021-04-05 2021-01-25            County Guilford     537174        Race
#> 71:  2021-04-05 2021-02-01            County Guilford     537174        Race
#> 72:  2021-04-05 2021-02-08            County Guilford     537174        Race
#> 73:  2021-04-05 2021-02-15            County Guilford     537174        Race
#> 74:  2021-04-05 2021-02-22            County Guilford     537174        Race
#> 75:  2021-04-05 2021-03-01            County Guilford     537174        Race
#> 76:  2021-04-05 2021-03-08            County Guilford     537174        Race
#> 77:  2021-04-05 2021-03-15            County Guilford     537174        Race
#> 78:  2021-04-05 2021-03-22            County Guilford     537174        Race
#> 79:  2021-04-05 2021-03-29            County Guilford     537174        Race
#> 80:  2021-04-05 2020-12-14            County Guilford     537174        Race
#> 81:  2021-04-05 2020-12-21            County Guilford     537174        Race
#> 82:  2021-04-05 2020-12-28            County Guilford     537174        Race
#> 83:  2021-04-05 2021-01-04            County Guilford     537174        Race
#> 84:  2021-04-05 2021-01-11            County Guilford     537174        Race
#> 85:  2021-04-05 2021-01-18            County Guilford     537174        Race
#> 86:  2021-04-05 2021-01-25            County Guilford     537174        Race
#> 87:  2021-04-05 2021-02-01            County Guilford     537174        Race
#> 88:  2021-04-05 2021-02-08            County Guilford     537174        Race
#> 89:  2021-04-05 2021-02-15            County Guilford     537174        Race
#> 90:  2021-04-05 2021-02-22            County Guilford     537174        Race
#> 91:  2021-04-05 2021-03-01            County Guilford     537174        Race
#> 92:  2021-04-05 2021-03-08            County Guilford     537174        Race
#> 93:  2021-04-05 2021-03-15            County Guilford     537174        Race
#> 94:  2021-04-05 2021-03-22            County Guilford     537174        Race
#> 95:  2021-04-05 2021-03-29            County Guilford     537174        Race
#> 96:  2021-04-05 2021-04-05            County Guilford     537174        Race
#>     date_pulled    week_of aggregation_level   county county_pop demographic
#>                  demographic_identity  status
#>  1: American Indian or Alaskan Native partial
#>  2: American Indian or Alaskan Native partial
#>  3: American Indian or Alaskan Native partial
#>  4: American Indian or Alaskan Native partial
#>  5: American Indian or Alaskan Native partial
#>  6: American Indian or Alaskan Native partial
#>  7: American Indian or Alaskan Native partial
#>  8: American Indian or Alaskan Native partial
#>  9: American Indian or Alaskan Native partial
#> 10: American Indian or Alaskan Native partial
#> 11: American Indian or Alaskan Native partial
#> 12: American Indian or Alaskan Native partial
#> 13: American Indian or Alaskan Native partial
#> 14: American Indian or Alaskan Native partial
#> 15: American Indian or Alaskan Native partial
#> 16: American Indian or Alaskan Native partial
#> 17: American Indian or Alaskan Native partial
#> 18:         Asian or Pacific Islander partial
#> 19:         Asian or Pacific Islander partial
#> 20:         Asian or Pacific Islander partial
#> 21:         Asian or Pacific Islander partial
#> 22:         Asian or Pacific Islander partial
#> 23:         Asian or Pacific Islander partial
#> 24:         Asian or Pacific Islander partial
#> 25:         Asian or Pacific Islander partial
#> 26:         Asian or Pacific Islander partial
#> 27:         Asian or Pacific Islander partial
#> 28:         Asian or Pacific Islander partial
#> 29:         Asian or Pacific Islander partial
#> 30:         Asian or Pacific Islander partial
#> 31:         Asian or Pacific Islander partial
#> 32:         Asian or Pacific Islander partial
#> 33:         Asian or Pacific Islander partial
#> 34:         Asian or Pacific Islander partial
#> 35:         Black or African American partial
#> 36:         Black or African American partial
#> 37:         Black or African American partial
#> 38:         Black or African American partial
#> 39:         Black or African American partial
#> 40:         Black or African American partial
#> 41:         Black or African American partial
#> 42:         Black or African American partial
#> 43:         Black or African American partial
#> 44:         Black or African American partial
#> 45:         Black or African American partial
#> 46:         Black or African American partial
#> 47:         Black or African American partial
#> 48:         Black or African American partial
#> 49:         Black or African American partial
#> 50:         Black or African American partial
#> 51:         Black or African American partial
#> 52:            Missing or Undisclosed partial
#> 53:            Missing or Undisclosed partial
#> 54:            Missing or Undisclosed partial
#> 55:            Missing or Undisclosed partial
#> 56:            Missing or Undisclosed partial
#> 57:            Missing or Undisclosed partial
#> 58:            Missing or Undisclosed partial
#> 59:            Missing or Undisclosed partial
#> 60:            Missing or Undisclosed partial
#> 61:            Missing or Undisclosed partial
#> 62:            Missing or Undisclosed partial
#> 63:            Missing or Undisclosed partial
#> 64:                             Other partial
#> 65:                             Other partial
#> 66:                             Other partial
#> 67:                             Other partial
#> 68:                             Other partial
#> 69:                             Other partial
#> 70:                             Other partial
#> 71:                             Other partial
#> 72:                             Other partial
#> 73:                             Other partial
#> 74:                             Other partial
#> 75:                             Other partial
#> 76:                             Other partial
#> 77:                             Other partial
#> 78:                             Other partial
#> 79:                             Other partial
#> 80:                             White partial
#> 81:                             White partial
#> 82:                             White partial
#> 83:                             White partial
#> 84:                             White partial
#> 85:                             White partial
#> 86:                             White partial
#> 87:                             White partial
#> 88:                             White partial
#> 89:                             White partial
#> 90:                             White partial
#> 91:                             White partial
#> 92:                             White partial
#> 93:                             White partial
#> 94:                             White partial
#> 95:                             White partial
#> 96:                             White partial
#>                  demographic_identity  status
#>                                 variable county_demo_pop vax_n
#>  1: People at Least Partially Vaccinated           23050     0
#>  2: People at Least Partially Vaccinated           23050    45
#>  3: People at Least Partially Vaccinated           23050    60
#>  4: People at Least Partially Vaccinated           23050    30
#>  5: People at Least Partially Vaccinated           23050    50
#>  6: People at Least Partially Vaccinated           23050   135
#>  7: People at Least Partially Vaccinated           23050   100
#>  8: People at Least Partially Vaccinated           23050   150
#>  9: People at Least Partially Vaccinated           23050   145
#> 10: People at Least Partially Vaccinated           23050    70
#> 11: People at Least Partially Vaccinated           23050   210
#> 12: People at Least Partially Vaccinated           23050   200
#> 13: People at Least Partially Vaccinated           23050   420
#> 14: People at Least Partially Vaccinated           23050   450
#> 15: People at Least Partially Vaccinated           23050   410
#> 16: People at Least Partially Vaccinated           23050   155
#> 17: People at Least Partially Vaccinated           23050     0
#> 18: People at Least Partially Vaccinated          149855   275
#> 19: People at Least Partially Vaccinated          149855   765
#> 20: People at Least Partially Vaccinated          149855   680
#> 21: People at Least Partially Vaccinated          149855   400
#> 22: People at Least Partially Vaccinated          149855   870
#> 23: People at Least Partially Vaccinated          149855  1440
#> 24: People at Least Partially Vaccinated          149855  1145
#> 25: People at Least Partially Vaccinated          149855  1235
#> 26: People at Least Partially Vaccinated          149855  1270
#> 27: People at Least Partially Vaccinated          149855   420
#> 28: People at Least Partially Vaccinated          149855  1575
#> 29: People at Least Partially Vaccinated          149855  2010
#> 30: People at Least Partially Vaccinated          149855  7540
#> 31: People at Least Partially Vaccinated          149855  7225
#> 32: People at Least Partially Vaccinated          149855  8240
#> 33: People at Least Partially Vaccinated          149855  2840
#> 34: People at Least Partially Vaccinated          149855     0
#> 35: People at Least Partially Vaccinated          982460   285
#> 36: People at Least Partially Vaccinated          982460  1490
#> 37: People at Least Partially Vaccinated          982460  1870
#> 38: People at Least Partially Vaccinated          982460  1795
#> 39: People at Least Partially Vaccinated          982460  4020
#> 40: People at Least Partially Vaccinated          982460 14190
#> 41: People at Least Partially Vaccinated          982460  5455
#> 42: People at Least Partially Vaccinated          982460 12050
#> 43: People at Least Partially Vaccinated          982460 10360
#> 44: People at Least Partially Vaccinated          982460  5560
#> 45: People at Least Partially Vaccinated          982460 20215
#> 46: People at Least Partially Vaccinated          982460 19730
#> 47: People at Least Partially Vaccinated          982460 29525
#> 48: People at Least Partially Vaccinated          982460 27390
#> 49: People at Least Partially Vaccinated          982460 29065
#> 50: People at Least Partially Vaccinated          982460 12490
#> 51: People at Least Partially Vaccinated          982460     0
#> 52: People at Least Partially Vaccinated              NA     5
#> 53: People at Least Partially Vaccinated              NA    15
#> 54: People at Least Partially Vaccinated              NA     5
#> 55: People at Least Partially Vaccinated              NA    25
#> 56: People at Least Partially Vaccinated              NA    40
#> 57: People at Least Partially Vaccinated              NA    95
#> 58: People at Least Partially Vaccinated              NA   435
#> 59: People at Least Partially Vaccinated              NA   720
#> 60: People at Least Partially Vaccinated              NA  1890
#> 61: People at Least Partially Vaccinated              NA  1565
#> 62: People at Least Partially Vaccinated              NA  1280
#> 63: People at Least Partially Vaccinated              NA   665
#> 64: People at Least Partially Vaccinated              NA   115
#> 65: People at Least Partially Vaccinated              NA   385
#> 66: People at Least Partially Vaccinated              NA   265
#> 67: People at Least Partially Vaccinated              NA   320
#> 68: People at Least Partially Vaccinated              NA   525
#> 69: People at Least Partially Vaccinated              NA  1625
#> 70: People at Least Partially Vaccinated              NA   855
#> 71: People at Least Partially Vaccinated              NA  1860
#> 72: People at Least Partially Vaccinated              NA  1965
#> 73: People at Least Partially Vaccinated              NA  1220
#> 74: People at Least Partially Vaccinated              NA  3235
#> 75: People at Least Partially Vaccinated              NA  3655
#> 76: People at Least Partially Vaccinated              NA  8965
#> 77: People at Least Partially Vaccinated              NA 11420
#> 78: People at Least Partially Vaccinated              NA 11480
#> 79: People at Least Partially Vaccinated              NA  4735
#> 80: People at Least Partially Vaccinated         1530505  2325
#> 81: People at Least Partially Vaccinated         1530505  7040
#> 82: People at Least Partially Vaccinated         1530505  6980
#> 83: People at Least Partially Vaccinated         1530505  8955
#> 84: People at Least Partially Vaccinated         1530505 24395
#> 85: People at Least Partially Vaccinated         1530505 56325
#> 86: People at Least Partially Vaccinated         1530505 23565
#> 87: People at Least Partially Vaccinated         1530505 37265
#> 88: People at Least Partially Vaccinated         1530505 24820
#> 89: People at Least Partially Vaccinated         1530505 11840
#> 90: People at Least Partially Vaccinated         1530505 37335
#> 91: People at Least Partially Vaccinated         1530505 35285
#> 92: People at Least Partially Vaccinated         1530505 56375
#> 93: People at Least Partially Vaccinated         1530505 55095
#> 94: People at Least Partially Vaccinated         1530505 51315
#> 95: People at Least Partially Vaccinated         1530505 21230
#> 96: People at Least Partially Vaccinated         1530505     0
#>                                 variable county_demo_pop vax_n