Pull Vaccinations by County and Demographic

Pulled from <conedatascience/covid-data>. The column vax_n returns vaccination counts. Population references are available for the user to calculate distribution and/or per capita rates.

get_vaccinations_demo(county_list = NULL, demographic = NULL, status = NULL)

Arguments

county_list

a string, the county to pull. default is all counties.

demographic

a string, the demographic requested. one or more of "race", "age", "ethnicity", "sex". default is all demographics

status

string, either 'partial' or 'full' for number of people with partially or fully vaccinated. Default of NULL returns both.

Examples

{

get_vaccinations_demo(county_list = "Guilford",
                      demographic = 'race', status = 'partial')

}
#>      date_pulled    week_of aggregation_level   county demographic
#>   1:  2021-12-19 2021-05-03            County Guilford        Race
#>   2:  2021-12-19 2021-05-24            County Guilford        Race
#>   3:  2021-12-19 2021-06-07            County Guilford        Race
#>   4:  2021-12-19 2021-06-28            County Guilford        Race
#>   5:  2021-12-19 2021-07-05            County Guilford        Race
#>  ---                                                              
#> 316:  2021-12-19 2021-11-29            County Guilford        Race
#> 317:  2021-12-19 2021-12-06            County Guilford        Race
#> 318:  2021-12-19 2021-12-13            County Guilford        Race
#> 319:  2021-12-19 2020-12-14            County Guilford        Race
#> 320:  2021-12-19 2021-01-04            County Guilford        Race
#>           demographic_identity  status                             variable
#>   1: Black or African American partial People at Least Partially Vaccinated
#>   2: Black or African American partial People at Least Partially Vaccinated
#>   3: Black or African American partial People at Least Partially Vaccinated
#>   4: Black or African American partial People at Least Partially Vaccinated
#>   5: Black or African American partial People at Least Partially Vaccinated
#>  ---                                                                       
#> 316:                     Other partial People at Least Partially Vaccinated
#> 317:                     Other partial People at Least Partially Vaccinated
#> 318:                     Other partial People at Least Partially Vaccinated
#> 319:                Suppressed partial People at Least Partially Vaccinated
#> 320:                Suppressed partial People at Least Partially Vaccinated
#>      county_demo_pop vax_n
#>   1:          196492  1461
#>   2:          196492  1565
#>   3:          196492  1026
#>   4:          196492   852
#>   5:          196492   779
#>  ---                      
#> 316:              NA   178
#> 317:              NA   227
#> 318:              NA    95
#> 319:              NA     0
#> 320:              NA     0