Pull Vaccination Rates by County and Demographic

Pulled from <conedatascience/covid-data>. The columns vax_prop and vax_n return the vaccination proportion and vaccination counts.

get_vaccinations_demo(
  county_list = NULL,
  measure = NULL,
  demographic = NULL,
  dose = NULL,
  latest = FALSE
)

Arguments

county_list
a string, the county to pull. default is all counties AND a North Carolina roll up. "North Carolina" can be supplied as a county.

measure
a string, the measure requested. one of "distribution" or "per capita" for pulling the distribution of vaccine (adds to 1 within the demographic) or the vaccinate rate per capita within the demographic, respectively. default is all measures.

demographic
a string, the demographic requested. one or more of "race", "age", "ethnicity", "sex". default is all demographics

dose
integer, the dose to pull (1 or 2). default is all doses.

latest
logical, whether to pull only the latest vaccination record or a full time series. default F pulls full time series

Examples

{

get_vaccinations_demo(county_list = "Guilford", measure = 'distribution',
                      demographic = 'race', dose = 1, latest = TRUE)

}

#>          date aggregation_level   county demographic
#> 1: 2021-03-03            County Guilford        Race
#> 2: 2021-03-03            County Guilford        Race
#> 3: 2021-03-03            County Guilford        Race
#> 4: 2021-03-03            County Guilford        Race
#> 5: 2021-03-03            County Guilford        Race
#> 6: 2021-03-03            County Guilford        Race
#>                 demographic_identity      measure dose
#> 1: American Indian or Alaskan Native Distribution    1
#> 2:         Asian or Pacific Islander Distribution    1
#> 3:         Black or African-American Distribution    1
#> 4:            Missing or Undisclosed Distribution    1
#> 5:                             Other Distribution    1
#> 6:                             White Distribution    1
#>                          variable    vax_prop vax_n demo_population
#> 1: First Doses ~ Percent of Total 0.002881886   199            4610
#> 2: First Doses ~ Percent of Total 0.029456062  2034           29971
#> 3: First Doses ~ Percent of Total 0.228074495 15749          196492
#> 4: First Doses ~ Percent of Total 0.002215721   153              NA
#> 5: First Doses ~ Percent of Total 0.036450791  2517              NA
#> 6: First Doses ~ Percent of Total 0.700921045 48400          306101
#>    demo_total_dose_1 demo_total_dose_2
#> 1:             69052             39135
#> 2:             69052             39135
#> 3:             69052             39135
#> 4:             69052             39135
#> 5:             69052             39135
#> 6:             69052             39135