Pull Vaccinations by County and Demographic

Pulled from <conedatascience/covid-data>. The column vax_n returns vaccination counts. Population references are available for the user to calculate distribution and/or per capita rates.

get_vaccinations_demo(county_list = NULL, demographic = NULL, status = NULL)

Arguments

county_list
a string, the county to pull. default is all counties.

demographic
a string, the demographic requested. one or more of "race", "age", "ethnicity", "sex". default is all demographics

status
string, either 'partial' or 'full' for number of people with partially or fully vaccinated. Default of NULL returns both.

Examples

{

get_vaccinations_demo(county_list = "Guilford",
                      demographic = 'race', status = 'partial')

}

#>     date_pulled    week_of aggregation_level   county county_pop demographic
#>  1:  2021-03-20 2020-12-14            County Guilford     537174        Race
#>  2:  2021-03-20 2020-12-21            County Guilford     537174        Race
#>  3:  2021-03-20 2020-12-28            County Guilford     537174        Race
#>  4:  2021-03-20 2021-01-04            County Guilford     537174        Race
#>  5:  2021-03-20 2021-01-11            County Guilford     537174        Race
#>  6:  2021-03-20 2021-01-18            County Guilford     537174        Race
#>  7:  2021-03-20 2021-01-25            County Guilford     537174        Race
#>  8:  2021-03-20 2021-02-01            County Guilford     537174        Race
#>  9:  2021-03-20 2021-02-08            County Guilford     537174        Race
#> 10:  2021-03-20 2021-02-15            County Guilford     537174        Race
#> 11:  2021-03-20 2021-02-22            County Guilford     537174        Race
#> 12:  2021-03-20 2021-03-01            County Guilford     537174        Race
#> 13:  2021-03-20 2021-03-08            County Guilford     537174        Race
#> 14:  2021-03-20 2021-03-15            County Guilford     537174        Race
#> 15:  2021-03-20 2020-12-14            County Guilford     537174        Race
#> 16:  2021-03-20 2020-12-21            County Guilford     537174        Race
#> 17:  2021-03-20 2020-12-28            County Guilford     537174        Race
#> 18:  2021-03-20 2021-01-04            County Guilford     537174        Race
#> 19:  2021-03-20 2021-01-11            County Guilford     537174        Race
#> 20:  2021-03-20 2021-01-18            County Guilford     537174        Race
#> 21:  2021-03-20 2021-01-25            County Guilford     537174        Race
#> 22:  2021-03-20 2021-02-01            County Guilford     537174        Race
#> 23:  2021-03-20 2021-02-08            County Guilford     537174        Race
#> 24:  2021-03-20 2021-02-15            County Guilford     537174        Race
#> 25:  2021-03-20 2021-02-22            County Guilford     537174        Race
#> 26:  2021-03-20 2021-03-01            County Guilford     537174        Race
#> 27:  2021-03-20 2021-03-08            County Guilford     537174        Race
#> 28:  2021-03-20 2021-03-15            County Guilford     537174        Race
#> 29:  2021-03-20 2020-12-14            County Guilford     537174        Race
#> 30:  2021-03-20 2020-12-21            County Guilford     537174        Race
#> 31:  2021-03-20 2020-12-28            County Guilford     537174        Race
#> 32:  2021-03-20 2021-01-04            County Guilford     537174        Race
#> 33:  2021-03-20 2021-01-11            County Guilford     537174        Race
#> 34:  2021-03-20 2021-01-18            County Guilford     537174        Race
#> 35:  2021-03-20 2021-01-25            County Guilford     537174        Race
#> 36:  2021-03-20 2021-02-01            County Guilford     537174        Race
#> 37:  2021-03-20 2021-02-08            County Guilford     537174        Race
#> 38:  2021-03-20 2021-02-15            County Guilford     537174        Race
#> 39:  2021-03-20 2021-02-22            County Guilford     537174        Race
#> 40:  2021-03-20 2021-03-01            County Guilford     537174        Race
#> 41:  2021-03-20 2021-03-08            County Guilford     537174        Race
#> 42:  2021-03-20 2021-03-15            County Guilford     537174        Race
#> 43:  2021-03-20 2020-12-28            County Guilford     537174        Race
#> 44:  2021-03-20 2021-01-18            County Guilford     537174        Race
#> 45:  2021-03-20 2021-01-25            County Guilford     537174        Race
#> 46:  2021-03-20 2021-02-01            County Guilford     537174        Race
#> 47:  2021-03-20 2021-02-08            County Guilford     537174        Race
#> 48:  2021-03-20 2021-02-15            County Guilford     537174        Race
#> 49:  2021-03-20 2021-02-22            County Guilford     537174        Race
#> 50:  2021-03-20 2021-03-01            County Guilford     537174        Race
#> 51:  2021-03-20 2021-03-08            County Guilford     537174        Race
#> 52:  2021-03-20 2021-03-15            County Guilford     537174        Race
#> 53:  2021-03-20 2020-12-14            County Guilford     537174        Race
#> 54:  2021-03-20 2020-12-21            County Guilford     537174        Race
#> 55:  2021-03-20 2020-12-28            County Guilford     537174        Race
#> 56:  2021-03-20 2021-01-04            County Guilford     537174        Race
#> 57:  2021-03-20 2021-01-11            County Guilford     537174        Race
#> 58:  2021-03-20 2021-01-18            County Guilford     537174        Race
#> 59:  2021-03-20 2021-01-25            County Guilford     537174        Race
#> 60:  2021-03-20 2021-02-01            County Guilford     537174        Race
#> 61:  2021-03-20 2021-02-08            County Guilford     537174        Race
#> 62:  2021-03-20 2021-02-15            County Guilford     537174        Race
#> 63:  2021-03-20 2021-02-22            County Guilford     537174        Race
#> 64:  2021-03-20 2021-03-01            County Guilford     537174        Race
#> 65:  2021-03-20 2021-03-08            County Guilford     537174        Race
#> 66:  2021-03-20 2021-03-15            County Guilford     537174        Race
#> 67:  2021-03-20 2020-12-14            County Guilford     537174        Race
#> 68:  2021-03-20 2020-12-21            County Guilford     537174        Race
#> 69:  2021-03-20 2020-12-28            County Guilford     537174        Race
#> 70:  2021-03-20 2021-01-04            County Guilford     537174        Race
#> 71:  2021-03-20 2021-01-11            County Guilford     537174        Race
#> 72:  2021-03-20 2021-01-18            County Guilford     537174        Race
#> 73:  2021-03-20 2021-01-25            County Guilford     537174        Race
#> 74:  2021-03-20 2021-02-01            County Guilford     537174        Race
#> 75:  2021-03-20 2021-02-08            County Guilford     537174        Race
#> 76:  2021-03-20 2021-02-15            County Guilford     537174        Race
#> 77:  2021-03-20 2021-02-22            County Guilford     537174        Race
#> 78:  2021-03-20 2021-03-01            County Guilford     537174        Race
#> 79:  2021-03-20 2021-03-08            County Guilford     537174        Race
#> 80:  2021-03-20 2021-03-15            County Guilford     537174        Race
#>     date_pulled    week_of aggregation_level   county county_pop demographic
#>                  demographic_identity  status
#>  1: American Indian or Alaskan Native partial
#>  2: American Indian or Alaskan Native partial
#>  3: American Indian or Alaskan Native partial
#>  4: American Indian or Alaskan Native partial
#>  5: American Indian or Alaskan Native partial
#>  6: American Indian or Alaskan Native partial
#>  7: American Indian or Alaskan Native partial
#>  8: American Indian or Alaskan Native partial
#>  9: American Indian or Alaskan Native partial
#> 10: American Indian or Alaskan Native partial
#> 11: American Indian or Alaskan Native partial
#> 12: American Indian or Alaskan Native partial
#> 13: American Indian or Alaskan Native partial
#> 14: American Indian or Alaskan Native partial
#> 15:         Asian or Pacific Islander partial
#> 16:         Asian or Pacific Islander partial
#> 17:         Asian or Pacific Islander partial
#> 18:         Asian or Pacific Islander partial
#> 19:         Asian or Pacific Islander partial
#> 20:         Asian or Pacific Islander partial
#> 21:         Asian or Pacific Islander partial
#> 22:         Asian or Pacific Islander partial
#> 23:         Asian or Pacific Islander partial
#> 24:         Asian or Pacific Islander partial
#> 25:         Asian or Pacific Islander partial
#> 26:         Asian or Pacific Islander partial
#> 27:         Asian or Pacific Islander partial
#> 28:         Asian or Pacific Islander partial
#> 29:         Black or African American partial
#> 30:         Black or African American partial
#> 31:         Black or African American partial
#> 32:         Black or African American partial
#> 33:         Black or African American partial
#> 34:         Black or African American partial
#> 35:         Black or African American partial
#> 36:         Black or African American partial
#> 37:         Black or African American partial
#> 38:         Black or African American partial
#> 39:         Black or African American partial
#> 40:         Black or African American partial
#> 41:         Black or African American partial
#> 42:         Black or African American partial
#> 43:            Missing or Undisclosed partial
#> 44:            Missing or Undisclosed partial
#> 45:            Missing or Undisclosed partial
#> 46:            Missing or Undisclosed partial
#> 47:            Missing or Undisclosed partial
#> 48:            Missing or Undisclosed partial
#> 49:            Missing or Undisclosed partial
#> 50:            Missing or Undisclosed partial
#> 51:            Missing or Undisclosed partial
#> 52:            Missing or Undisclosed partial
#> 53:                             Other partial
#> 54:                             Other partial
#> 55:                             Other partial
#> 56:                             Other partial
#> 57:                             Other partial
#> 58:                             Other partial
#> 59:                             Other partial
#> 60:                             Other partial
#> 61:                             Other partial
#> 62:                             Other partial
#> 63:                             Other partial
#> 64:                             Other partial
#> 65:                             Other partial
#> 66:                             Other partial
#> 67:                             White partial
#> 68:                             White partial
#> 69:                             White partial
#> 70:                             White partial
#> 71:                             White partial
#> 72:                             White partial
#> 73:                             White partial
#> 74:                             White partial
#> 75:                             White partial
#> 76:                             White partial
#> 77:                             White partial
#> 78:                             White partial
#> 79:                             White partial
#> 80:                             White partial
#>                  demographic_identity  status
#>                                 variable county_demo_pop vax_n
#>  1: People at Least Partially Vaccinated            4610     0
#>  2: People at Least Partially Vaccinated            4610     9
#>  3: People at Least Partially Vaccinated            4610    12
#>  4: People at Least Partially Vaccinated            4610     6
#>  5: People at Least Partially Vaccinated            4610    10
#>  6: People at Least Partially Vaccinated            4610    27
#>  7: People at Least Partially Vaccinated            4610    20
#>  8: People at Least Partially Vaccinated            4610    30
#>  9: People at Least Partially Vaccinated            4610    29
#> 10: People at Least Partially Vaccinated            4610    14
#> 11: People at Least Partially Vaccinated            4610    42
#> 12: People at Least Partially Vaccinated            4610    40
#> 13: People at Least Partially Vaccinated            4610    83
#> 14: People at Least Partially Vaccinated            4610    71
#> 15: People at Least Partially Vaccinated           29971    55
#> 16: People at Least Partially Vaccinated           29971   153
#> 17: People at Least Partially Vaccinated           29971   136
#> 18: People at Least Partially Vaccinated           29971    80
#> 19: People at Least Partially Vaccinated           29971   174
#> 20: People at Least Partially Vaccinated           29971   286
#> 21: People at Least Partially Vaccinated           29971   226
#> 22: People at Least Partially Vaccinated           29971   247
#> 23: People at Least Partially Vaccinated           29971   253
#> 24: People at Least Partially Vaccinated           29971    83
#> 25: People at Least Partially Vaccinated           29971   313
#> 26: People at Least Partially Vaccinated           29971   402
#> 27: People at Least Partially Vaccinated           29971  1498
#> 28: People at Least Partially Vaccinated           29971  1074
#> 29: People at Least Partially Vaccinated          196492    57
#> 30: People at Least Partially Vaccinated          196492   298
#> 31: People at Least Partially Vaccinated          196492   374
#> 32: People at Least Partially Vaccinated          196492   359
#> 33: People at Least Partially Vaccinated          196492   801
#> 34: People at Least Partially Vaccinated          196492  2831
#> 35: People at Least Partially Vaccinated          196492  1089
#> 36: People at Least Partially Vaccinated          196492  2410
#> 37: People at Least Partially Vaccinated          196492  2070
#> 38: People at Least Partially Vaccinated          196492  1109
#> 39: People at Least Partially Vaccinated          196492  4036
#> 40: People at Least Partially Vaccinated          196492  3922
#> 41: People at Least Partially Vaccinated          196492  5869
#> 42: People at Least Partially Vaccinated          196492  4301
#> 43: People at Least Partially Vaccinated              NA     1
#> 44: People at Least Partially Vaccinated              NA     3
#> 45: People at Least Partially Vaccinated              NA     1
#> 46: People at Least Partially Vaccinated              NA     4
#> 47: People at Least Partially Vaccinated              NA     7
#> 48: People at Least Partially Vaccinated              NA    19
#> 49: People at Least Partially Vaccinated              NA    86
#> 50: People at Least Partially Vaccinated              NA   142
#> 51: People at Least Partially Vaccinated              NA   372
#> 52: People at Least Partially Vaccinated              NA   265
#> 53: People at Least Partially Vaccinated              NA    23
#> 54: People at Least Partially Vaccinated              NA    77
#> 55: People at Least Partially Vaccinated              NA    53
#> 56: People at Least Partially Vaccinated              NA    64
#> 57: People at Least Partially Vaccinated              NA   105
#> 58: People at Least Partially Vaccinated              NA   322
#> 59: People at Least Partially Vaccinated              NA   170
#> 60: People at Least Partially Vaccinated              NA   372
#> 61: People at Least Partially Vaccinated              NA   392
#> 62: People at Least Partially Vaccinated              NA   243
#> 63: People at Least Partially Vaccinated              NA   642
#> 64: People at Least Partially Vaccinated              NA   729
#> 65: People at Least Partially Vaccinated              NA  1788
#> 66: People at Least Partially Vaccinated              NA  1629
#> 67: People at Least Partially Vaccinated          306101   465
#> 68: People at Least Partially Vaccinated          306101  1408
#> 69: People at Least Partially Vaccinated          306101  1396
#> 70: People at Least Partially Vaccinated          306101  1790
#> 71: People at Least Partially Vaccinated          306101  4841
#> 72: People at Least Partially Vaccinated          306101 11216
#> 73: People at Least Partially Vaccinated          306101  4709
#> 74: People at Least Partially Vaccinated          306101  7448
#> 75: People at Least Partially Vaccinated          306101  4953
#> 76: People at Least Partially Vaccinated          306101  2354
#> 77: People at Least Partially Vaccinated          306101  7463
#> 78: People at Least Partially Vaccinated          306101  7031
#> 79: People at Least Partially Vaccinated          306101 11192
#> 80: People at Least Partially Vaccinated          306101  8657
#>                                 variable county_demo_pop vax_n