Pull Vaccinations by County and Demographic

Pulled from <conedatascience/covid-data>. The column vax_n returns vaccination counts. Population references are available for the user to calculate distribution and/or per capita rates.

get_vaccinations_demo(county_list = NULL, demographic = NULL, status = NULL)

Arguments

county_list
a string, the county to pull. default is all counties.

demographic
a string, the demographic requested. one or more of "race", "age", "ethnicity", "sex". default is all demographics

status
string, either 'partial' or 'full' for number of people with partially or fully vaccinated. Default of NULL returns both.

Examples

{

get_vaccinations_demo(county_list = "Guilford",
                      demographic = 'race', status = 'partial')

}

#>      date_pulled    week_of aggregation_level   county county_pop demographic
#>   1:  2021-08-02 2021-05-24            County Guilford       1988        Race
#>   2:  2021-08-02 2021-03-29            County Guilford      10424        Race
#>   3:  2021-08-02 2021-03-01            County Guilford      12497        Race
#>   4:  2021-08-02 2020-12-14            County Guilford        600        Race
#>   5:  2021-08-02 2020-12-21            County Guilford       1944        Race
#>  ---                                                                         
#> 194:  2021-08-02 2021-06-28            County Guilford        655        Race
#> 195:  2021-08-02 2021-07-05            County Guilford        614        Race
#> 196:  2021-08-02 2021-07-12            County Guilford        753        Race
#> 197:  2021-08-02 2021-07-19            County Guilford        957        Race
#> 198:  2021-08-02 2021-07-26            County Guilford       1209        Race
#>                   demographic_identity  status
#>   1:         Black or African American partial
#>   2:         Black or African American partial
#>   3:                             White partial
#>   4: American Indian or Alaskan Native partial
#>   5: American Indian or Alaskan Native partial
#>  ---                                          
#> 194:                             White partial
#> 195:                             White partial
#> 196:                             White partial
#> 197:                             White partial
#> 198:                             White partial
#>                                  variable county_demo_pop vax_n
#>   1: People at Least Partially Vaccinated          196492  1550
#>   2: People at Least Partially Vaccinated          196492  3905
#>   3: People at Least Partially Vaccinated          306101 10493
#>   4: People at Least Partially Vaccinated            4610     0
#>   5: People at Least Partially Vaccinated            4610     9
#>  ---                                                           
#> 194: People at Least Partially Vaccinated          306101   726
#> 195: People at Least Partially Vaccinated          306101   708
#> 196: People at Least Partially Vaccinated          306101   836
#> 197: People at Least Partially Vaccinated          306101  1264
#> 198: People at Least Partially Vaccinated          306101  1226